requirements.txt
# YOLO Combined Model Project Requirements
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Web UI
streamlit>=1.28.0
plotly>=5.17.0

# Optional for advanced features
tensorboard>=2.13.0

# ONNX support (optional, uncomment if needed and compatible with your Python version)
# onnx>=1.14.0
# onnxruntime>=1.15.0
